move testproject under testprojects
