+++ testprojects/src/java/org/pantsbuild/testproject/runner/DependentClass.java
@@ -1,7 +1,7 @@
 // Copyright 2015 Pants project contributors (see CONTRIBUTORS.md).
 // Licensed under the Apache License, Version 2.0 (see LICENSE).
 
-package org.pantsbuild.tools.runner.testproject;
+package org.pantsbuild.testproject.runner;
 
 public class DependentClass {
   public static String getDependentClassMessage() {
