+++ testprojects/src/java/org/pantsbuild/testproject/runner/MainClass.java
@@ -1,7 +1,7 @@
 // Copyright 2015 Pants project contributors (see CONTRIBUTORS.md).
 // Licensed under the Apache License, Version 2.0 (see LICENSE).
 
-package org.pantsbuild.tools.runner.testproject;
+package org.pantsbuild.testproject.runner;
 
 import java.net.URL;
 import java.net.URLClassLoader;
